+++ tests/testthat/test-scatterplotmatrix_app.R
@@ -18,7 +18,7 @@ ID <- poc( # nolint
   )
 )
 
-root_app <- start_app_driver(dv.biomarker.general::mock_app_scatterplotmatrix())
+root_app <- start_app_driver(dv.explorer.parameter::mock_app_scatterplotmatrix())
 on.exit(if ("stop" %in% names(root_app)) root_app$stop())
 
 fail_if_app_not_started <- function() {
@@ -145,7 +145,7 @@ test_that("default values are set", {
 
   app <- start_app_driver(
     rlang::quo(
-      dv.biomarker.general::mock_app_scatterplotmatrix(
+      dv.explorer.parameter::mock_app_scatterplotmatrix(
         srv_defaults = !!srv_defaults
       )
     )