+++ NEWS.md
@@ -1,3 +1,9 @@
+# dv.explorer.parameter 0.0.5
+
+* First Github release
+* Renamed to dv.explorer.parameter
+
+
 # dv.biomarker.general 0.0.4
 
 * Correlation heatmap: